[[wip]]: decompose state machine
