+++ node/src/node/slave/state/active.cpp
@@ -1,6 +1,7 @@
 #include "cocaine/detail/service/node/slave/state/active.hpp"
 
 #include "cocaine/service/node/slave.hpp"
+#include "cocaine/detail/service/node/slave/machine.hpp"
 
 #include "cocaine/detail/service/node/slave/control.hpp"
 #include "cocaine/detail/service/node/slave/state/sealing.hpp"
