+++ node/src/node/slave/state/sealing.cpp
@@ -3,6 +3,7 @@
 #include <blackhole/logger.hpp>
 
 #include "cocaine/service/node/slave.hpp"
+#include "cocaine/detail/service/node/slave/machine.hpp"
 
 #include "cocaine/detail/service/node/slave/state/terminating.hpp"
